Module hvif_rs::types

Types for HVIF images

Reexports

pub use self::style::*;

pub use self::path::*;

pub use self::shape::*;

Modules

path
Types for HVIF paths

shape
Types for HVIF shapes

style
Types for HVIF styles

Structs

HVIFFlag
Flags that modify the parsing of the following data

HVIFImage
A fully specified HVIF image

Constants

HVIF_GRADIENT_FLAG_COLORS_16_BIT
The colors in this gradient have a deeper color depth per channel of 16 bits, instead of 8

HVIF_GRADIENT_FLAG_GRAYS
The colors in this gradient are all grayscale

HVIF_GRADIENT_FLAG_NO_ALPHA
The colors in this gradient have no alpha channel

HVIF_GRADIENT_FLAG_TRANSFORM
?

HVIF_PATH_FLAG_CLOSED
The path's last point is connected to its first point

HVIF_PATH_FLAG_NO_CURVES
The path is made up entirely of straight lines

HVIF_PATH_FLAG_USES_COMMANDS
The path has a command section, and can use HLine and VLine commands

HVIF_SHAPE_FLAG_HAS_TRANSFORMERS
?

HVIF_SHAPE_FLAG_HINTING
?

HVIF_SHAPE_FLAG_LOD_SCALE
?

HVIF_SHAPE_FLAG_TRANSFORM
This shape uses an affine transformation matrix to translate, rotate, and scale itself

HVIF_SHAPE_FLAG_TRANSLATION
?