Struct hvif_rs::types::style::HVIFGradientColor

pub struct HVIFGradientColor {
    pub stop_offset: u8,
    pub alpha: u8,
    pub red: u8,
    pub green: u8,
    pub blue: u8,
}

Color of a gradient, along with a stop offset

Fields
stop_offset: u8

The stop offset for this color - the "position" it holds relative to other colors in the gradient

alpha: u8

The alpha channel

red: u8

The red channel

green: u8

The green channel

blue: u8

The blue channel